backend/python/requirements.txt
fastapi
mysql-connector-python
uvicorn
orjson